file_service/app/requirements.txt
grpcio
grpcio-tools
psycopg2-binary
sqlmodel